move preprocess to utils
