+++ utils.py
@@ -39,3 +39,11 @@ def read_files(file_names, open_file_function=smart_open):
         with open_file_function(name) as f:
             for line in f:
                 yield line if isinstance(line, str) else line.decode('utf8')
+
+
+def preprocess(stack_trace):
+    def clean(func):
+        func = func.lower().replace('\n', '')
+        return func[:func.index('@0x') + 3] if '@0x' in func else func
+
+    return [clean(f) for f in stack_trace.split(' | ')]
